convert draggable utils to typescript
